backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
google-cloud-vision==3.4.4
google-cloud-storage==2.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2 